requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
motor==3.4.0
pymongo==4.6.2
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
requests
python-multipart==0.0.20
backports.zoneinfo==0.2.1; python_version < "3.9"